[Component] Address base component interface feedback #644
* Removed currently unused classes and attributes properties
* Moved background style constants and getter to Container
