+++ bundles/core/src/main/java/com/adobe/cq/wcm/core/components/models/Container.java
@@ -19,6 +19,7 @@ import java.util.List;
 import java.util.Map;
 
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 import org.osgi.annotation.versioning.ConsumerType;
 
 import com.adobe.cq.export.json.ComponentExporter;
@@ -33,6 +34,41 @@ import com.adobe.cq.export.json.ContainerExporter;
 public interface Container extends Component, ContainerExporter {
 
     /**
+     * Name of the configuration policy property that indicates if background images are enabled
+     *
+     * @since com.adobe.cq.wcm.core.components.models 12.8.0
+     */
+    String PN_BACKGROUND_IMAGE_ENABLED = "backgroundImageEnabled";
+
+    /**
+     * Name of the configuration policy property that indicates if background colors are enabled
+     *
+     * @since com.adobe.cq.wcm.core.components.models 12.8.0
+     */
+    String PN_BACKGROUND_COLOR_ENABLED = "backgroundColorEnabled";
+
+    /**
+     * Name of the configuration policy property that indicates if background colors are to be restricted to predefined values
+     *
+     * @since com.adobe.cq.wcm.core.components.models 12.8.0
+     */
+    String PN_BACKGROUND_COLOR_SWATCHES_ONLY = "backgroundColorSwatchesOnly";
+
+    /**
+     * Name of the resource property that indicates that path to the background image
+     *
+     * @since com.adobe.cq.wcm.core.components.models 12.8.0
+     */
+    String PN_BACKGROUND_IMAGE_REFERENCE = "backgroundImageReference";
+
+    /**
+     * Name of the resource property that indicates the background color
+     *
+     * @since com.adobe.cq.wcm.core.components.models 12.8.0
+     */
+    String PN_BACKGROUND_COLOR = "backgroundColor";
+
+    /**
      * Returns a list of container items
      *
      * @return List of container items
@@ -40,6 +76,17 @@ public interface Container extends Component, ContainerExporter {
      */
     @NotNull
     default List<ListItem> getItems() {
+        throw new UnsupportedOperationException();
+    }
+
+    /**
+     * Returns the background CSS style to be applied to the component's root element
+     *
+     * @return CSS style string for the component's root element
+     * @since com.adobe.cq.wcm.core.components.models 12.8.0
+     */
+    @Nullable
+    default String getBackgroundStyle() {
         throw new UnsupportedOperationException();
     }
