+++ content/src/content/jcr_root/apps/core/wcm/components/container/v1/container/responsiveGrid.html
@@ -16,7 +16,7 @@
 <template data-sly-template.responsiveGrid="${ @ container}">
     <div id="${container.id}"
          class="cmp-container"
-         style="${container.style @ context='styleString'}">
+         style="${container.backgroundStyle @ context='styleString'}">
         <sly data-sly-resource="${resource @ resourceType='wcm/foundation/components/responsivegrid'}"></sly>
     </div>
 </template>
